Use RegExp for matching position.
Now and instead, match against a simple regular expression that matches
a number of potential cases. Data entry can now give us any value it
wants to that conforms to a very simple standard (starts with "r" or
"b", ends in a number) and we can likely parse it.

It is, of course, preferred to give such simple values as "r1", "red1",
or "red_1", but the system can account for that now.

I also added a lot of comments. @Sammidysam was right, I like to think I
*did* improve it.