Add hub files to load controllers and models.
I do this (a) because it allows us to require "frecon/models" instead of
looping through all of the files in the models directory and (b) because
if someone were to put an experimental model in there that they didn't
want getting loaded in the server (Yes, really obscure use case, I
know), that would cause problems. This is simply cleaner to me.