Add Model class that all models extend
WARNING: This makes your current database obsolete.  The system will not
pick up the data in it anymore.  This is because the :query parameter
changed from {} to {"_type"=>{"$in"=>["FReCon::Team"]}} as a result of
this change.  Since we are not yet at version 1.0, this change can be
made.  It will help reduce repetitive code in the future.