FReCon::Database.setup: Enable custom Hash for configuration.
If a Hash is given, it is converted to yaml (via to_yaml) and written to a Tempfile. This Tempfile's path is passed to the Mongoid.load! method instead of the path to the "mongoid.yml" file.  By default (If no `mongoid_hash' is given), FReCon::Database.setup will give Mongoid.load! the "mongoid.yml" file in the "lib/frecon/" directory's path.