Remove attr_readers, use nil default context
We do not need the attr_readers because outside of the class the return_value
method is called which gets the values internally.  A nil default context is
more concise than defaulting to an error message because if someone sees
"context": null, they will understand that there is no context.