Rename base_url to base_uri
I figure I might have to argue my position on this one, so here you go:

Functionally, everything is just the same, as we aren't using either
yet.  The idea is that URLs are a subset of URIs, and since we aren't
performing validations, technically its safer to say that the instance
variables are URIs.

I say this because some people like to use shortnames like "google.com"
instead of "https://www.google.com".  The former is a URI and the latter
is a URL.  Some (specifically us) who write more Scrapers in the future
might feel compelled to use such a shorthand without the specific
protocol to be used, and so it's semantically safer to call things the
URI.

RFC 3986 defines URIs and URLs as follows:

"A Uniform Resource Identifier (URI) is a compact sequence of characters
that identifies an abstract or physical resource."

"A URI can be further classified as a locator, a name, or both. The term
“Uniform Resource Locator” (URL) refers to the subset of URIs that, in
addition to identifying a resource, provide a means of locating the
resource by describing its primary access mechanism (e.g., its network
“location”)."

Further reading: https://danielmiessler.com/study/url_vs_uri/

Also, Ruby's Net::HTTP offers the ability to create requests using URIs
as destination designators.  And HTTParty uses the term "URI"
internally, as well.  It just makes sense to make everything all
agreeable.