Require "frecon/request_error."
This is another instance of where we probably should be requiring a file
but we aren't.  Again, it never hurts to require something
superfluously, but it'd be really great if we could require all of the
files in a project without any worries as to whether or not stuff is
already defined.

This lib/frecon/controller.rb file---and, really, all other files in the
project---should explicitly define their dependencies.  Since we were
using RequestError here, I decided to go ahead and require it.  In this
way, we can now specifically require "frecon/controller" and test each
of the various controller methods without worrying about other stuff.