Attempt to solve certain Travis issues
It appears that the usage of the MongoDB service requires non-container
infrastructure on Travis.  Travis is attempting to deprecate this, and
the container infrastructure affords much lower push-to-reaction times;
containers can have cached bundles, etc.

Since we've got the option, we can simply mock out database
functionality within our spec, once we get around to it.  This will not
only make things cleaner, but it will also make testing easier and
faster, and it only makes sense.