frecon/console: Remove preredundant require for frecon/database
Since we're later requiring `frecon`, which is a qualified "hub file" per [GitHub] #103, we don't need to require it again.

I think the general purpose of frecon/console was to act as an interface into everything that was loaded by the `frecon` file anyways, so this simply removes the other require.