match_number_spec: Switch to dynamic metaprogramming spec layout
This is much better and will ensure full coverage as more edge cases are
tested.  Note that I did switch it so that it actually tests for
each individual case on a by-method basis rather than by-method on an
individual case basis.  I might write a bootstrap method which builds
the context and such within the describe blocks.  This is a good first
step, however.