Update Competition model to resolve build errors
It appears that Mongoid 6.0 brings new breaking changes for us.  We can no longer use the name `matches?' as a method (which is automagically created when we define a relation named matches) because that's already defined in the core of Mongoid.  I hope to open an issue soon regarding this new behavior.